+++ src/Lesson7/Main.java
@@ -1,4 +1,4 @@
-package Lesson7.Poker;
+package Lesson7;
 
 public class Main {
     public static void main(String[] args) {
@@ -25,4 +25,5 @@ public class Main {
         System.out.println(playerFour);
 
     }
+
 }
